+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 1.0.5
-*
+* Fix package is_installed method to handle packages that haven't been uploaded
 
 ### 1.0.4
 * Fix config property's operation in spec file